Use StringThings
Use String things instead of jamming it all in ConfBlock
